explicit conversions

+++ src/lib_utils/log.cpp
@@ -41,7 +41,7 @@ std::string Log::getTime() {
 	const std::time_t t = std::time(nullptr);
 	const std::tm tm = *std::localtime(&t);
 	auto const size = strftime(szOut, 255, "%Y/%m/%d %H:%M:%S", &tm);
-	return format("[%s][%s] ", std::string(szOut, size), g_DefaultClock->now());
+	return format("[%s][%s] ", std::string(szOut, size), (double)g_DefaultClock->now());
 }
 
 std::string Log::getColorBegin(Level level) {
